+++ mne/viz/_brain/_timeviewer.py
@@ -15,15 +15,17 @@ import warnings
 import numpy as np
 from scipy import sparse
 
-from . import _Brain
+from ._brain import _Brain
+from .callback import (ShowView, IntSlider, TimeSlider, SmartSlider,
+                       BumpColorbarPoints, UpdateColorbarScale)
+from .mplcanvas import MplCanvas
 from .view import _lh_views_dict
 
-from ..utils import _show_help, _get_color_list, tight_layout
+from ..utils import _show_help, _get_color_list
 from ...externals.decorator import decorator
 from ...source_space import vertex_to_mni, _read_talxfm
 from ...transforms import apply_trans
 from ...utils import _ReuseCycle, warn, copy_doc, _validate_type
-from ...fixes import nullcontext
 
 
 @decorator
@@ -33,293 +35,6 @@ def safe_event(fun, *args, **kwargs):
         return fun(*args, **kwargs)
     except Exception:
         traceback.print_exc(file=sys.stderr)
-
-
-class MplCanvas(object):
-    """Ultimately, this is a QWidget (as well as a FigureCanvasAgg, etc.)."""
-
-    def __init__(self, time_viewer, width, height, dpi):
-        from PyQt5 import QtWidgets
-        from matplotlib import rc_context
-        from matplotlib.figure import Figure
-        from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg
-        if time_viewer.separate_canvas:
-            parent = None
-        else:
-            parent = time_viewer.window
-        # prefer constrained layout here but live with tight_layout otherwise
-        context = nullcontext
-        extra_events = ('resize',)
-        try:
-            context = rc_context({'figure.constrained_layout.use': True})
-            extra_events = ()
-        except KeyError:
-            pass
-        with context:
-            self.fig = Figure(figsize=(width, height), dpi=dpi)
-        self.canvas = FigureCanvasQTAgg(self.fig)
-        self.axes = self.fig.add_subplot(111)
-        self.axes.set(xlabel='Time (sec)', ylabel='Activation (AU)')
-        self.canvas.setParent(parent)
-        FigureCanvasQTAgg.setSizePolicy(
-            self.canvas,
-            QtWidgets.QSizePolicy.Expanding,
-            QtWidgets.QSizePolicy.Expanding
-        )
-        FigureCanvasQTAgg.updateGeometry(self.canvas)
-        self.time_viewer = time_viewer
-        self.time_func = time_viewer.callbacks["time"]
-        for event in ('button_press', 'motion_notify') + extra_events:
-            self.canvas.mpl_connect(
-                event + '_event', getattr(self, 'on_' + event))
-
-    def plot(self, x, y, label, **kwargs):
-        """Plot a curve."""
-        line, = self.axes.plot(
-            x, y, label=label, **kwargs)
-        self.update_plot()
-        return line
-
-    def plot_time_line(self, x, label, **kwargs):
-        """Plot the vertical line."""
-        line = self.axes.axvline(x, label=label, **kwargs)
-        self.update_plot()
-        return line
-
-    def update_plot(self):
-        """Update the plot."""
-        leg = self.axes.legend(
-            prop={'family': 'monospace', 'size': 'small'},
-            framealpha=0.5, handlelength=1.,
-            facecolor=self.time_viewer.brain._bg_color)
-        for text in leg.get_texts():
-            text.set_color(self.time_viewer.brain._fg_color)
-        with warnings.catch_warnings(record=True):
-            warnings.filterwarnings('ignore', 'constrained_layout')
-            self.canvas.draw()
-
-    def set_color(self, bg_color, fg_color):
-        """Set the widget colors."""
-        self.axes.set_facecolor(bg_color)
-        self.axes.xaxis.label.set_color(fg_color)
-        self.axes.yaxis.label.set_color(fg_color)
-        self.axes.spines['top'].set_color(fg_color)
-        self.axes.spines['bottom'].set_color(fg_color)
-        self.axes.spines['left'].set_color(fg_color)
-        self.axes.spines['right'].set_color(fg_color)
-        self.axes.tick_params(axis='x', colors=fg_color)
-        self.axes.tick_params(axis='y', colors=fg_color)
-        self.fig.patch.set_facecolor(bg_color)
-
-    def show(self):
-        """Show the canvas."""
-        self.canvas.show()
-
-    def close(self):
-        """Close the canvas."""
-        self.canvas.close()
-
-    def on_button_press(self, event):
-        """Handle button presses."""
-        # left click (and maybe drag) in progress in axes
-        if (event.inaxes != self.axes or
-                event.button != 1):
-            return
-        self.time_func(
-            event.xdata, update_widget=True, time_as_index=False)
-
-    def clear(self):
-        """Clear internal variables."""
-        self.close()
-        self.axes.clear()
-        self.fig.clear()
-        self.time_viewer = None
-        self.canvas = None
-
-    on_motion_notify = on_button_press  # for now they can be the same
-
-    def on_resize(self, event):
-        """Handle resize events."""
-        tight_layout(fig=self.axes.figure)
-
-
-class IntSlider(object):
-    """Class to set a integer slider."""
-
-    def __init__(self, plotter=None, callback=None, first_call=True):
-        self.plotter = plotter
-        self.callback = callback
-        self.slider_rep = None
-        self.first_call = first_call
-        self._first_time = True
-
-    def __call__(self, value):
-        """Round the label of the slider."""
-        idx = int(round(value))
-        if self.slider_rep is not None:
-            self.slider_rep.SetValue(idx)
-            self.plotter.update()
-        if not self._first_time or all([self._first_time, self.first_call]):
-            self.callback(idx)
-        if self._first_time:
-            self._first_time = False
-
-
-class TimeSlider(object):
-    """Class to update the time slider."""
-
-    def __init__(self, plotter=None, brain=None, callback=None,
-                 first_call=True):
-        self.plotter = plotter
-        self.brain = brain
-        self.callback = callback
-        self.slider_rep = None
-        self.first_call = first_call
-        self._first_time = True
-        self.time_label = None
-        if self.brain is not None and callable(self.brain._data['time_label']):
-            self.time_label = self.brain._data['time_label']
-
-    def __call__(self, value, update_widget=False, time_as_index=True):
-        """Update the time slider."""
-        value = float(value)
-        if not time_as_index:
-            value = self.brain._to_time_index(value)
-        if not self._first_time or all([self._first_time, self.first_call]):
-            self.brain.set_time_point(value)
-        if self.callback is not None:
-            self.callback()
-        current_time = self.brain._current_time
-        if self.slider_rep is not None:
-            if self.time_label is not None:
-                current_time = self.time_label(current_time)
-                self.slider_rep.SetTitleText(current_time)
-            if update_widget:
-                self.slider_rep.SetValue(value)
-                self.plotter.update()
-        if self._first_time:
-            self._first_time = False
-
-
-class UpdateColorbarScale(object):
-    """Class to update the values of the colorbar sliders."""
-
-    def __init__(self, plotter=None, brain=None):
-        self.plotter = plotter
-        self.brain = brain
-        self.keys = ('fmin', 'fmid', 'fmax')
-        self.reps = {key: None for key in self.keys}
-        self.slider_rep = None
-
-    def __call__(self, value):
-        """Update the colorbar sliders."""
-        self.brain._update_fscale(value)
-        for key in self.keys:
-            if self.reps[key] is not None:
-                self.reps[key].SetValue(self.brain._data[key])
-        if self.slider_rep is not None:
-            self.slider_rep.SetValue(1.0)
-        self.plotter.update()
-
-
-class BumpColorbarPoints(object):
-    """Class that ensure constraints over the colorbar points."""
-
-    def __init__(self, plotter=None, brain=None, name=None):
-        self.plotter = plotter
-        self.brain = brain
-        self.name = name
-        self.callback = {
-            "fmin": lambda fmin: brain.update_lut(fmin=fmin),
-            "fmid": lambda fmid: brain.update_lut(fmid=fmid),
-            "fmax": lambda fmax: brain.update_lut(fmax=fmax),
-        }
-        self.keys = ('fmin', 'fmid', 'fmax')
-        self.reps = {key: None for key in self.keys}
-        self.last_update = time.time()
-
-    def __call__(self, value):
-        """Update the colorbar sliders."""
-        vals = {key: self.brain._data[key] for key in self.keys}
-        if self.name == "fmin" and self.reps["fmin"] is not None:
-            if vals['fmax'] < value:
-                vals['fmax'] = value
-                self.reps['fmax'].SetValue(value)
-            if vals['fmid'] < value:
-                vals['fmid'] = value
-                self.reps['fmid'].SetValue(value)
-            self.reps['fmin'].SetValue(value)
-        elif self.name == "fmid" and self.reps['fmid'] is not None:
-            if vals['fmin'] > value:
-                vals['fmin'] = value
-                self.reps['fmin'].SetValue(value)
-            if vals['fmax'] < value:
-                vals['fmax'] = value
-                self.reps['fmax'].SetValue(value)
-            self.reps['fmid'].SetValue(value)
-        elif self.name == "fmax" and self.reps['fmax'] is not None:
-            if vals['fmin'] > value:
-                vals['fmin'] = value
-                self.reps['fmin'].SetValue(value)
-            if vals['fmid'] > value:
-                vals['fmid'] = value
-                self.reps['fmid'].SetValue(value)
-            self.reps['fmax'].SetValue(value)
-        self.brain.update_lut(**vals)
-        if time.time() > self.last_update + 1. / 60.:
-            self.callback[self.name](value)
-            self.last_update = time.time()
-        self.plotter.update()
-
-
-class ShowView(object):
-    """Class that selects the correct view."""
-
-    def __init__(self, plotter=None, brain=None, orientation=None,
-                 row=None, col=None, hemi=None):
-        self.plotter = plotter
-        self.brain = brain
-        self.orientation = orientation
-        self.short_orientation = [s[:3] for s in orientation]
-        self.row = row
-        self.col = col
-        self.hemi = hemi
-        self.slider_rep = None
-
-    def __call__(self, value, update_widget=False):
-        """Update the view."""
-        self.brain.show_view(value, row=self.row, col=self.col,
-                             hemi=self.hemi)
-        if update_widget:
-            if len(value) > 3:
-                idx = self.orientation.index(value)
-            else:
-                idx = self.short_orientation.index(value)
-            if self.slider_rep is not None:
-                self.slider_rep.SetValue(idx)
-                self.slider_rep.SetTitleText(self.orientation[idx])
-                self.plotter.update()
-
-
-class SmartSlider(object):
-    """Class to manage smart slider.
-
-    It stores it's own slider representation for efficiency
-    and uses it when necessary.
-    """
-
-    def __init__(self, plotter=None, callback=None):
-        self.plotter = plotter
-        self.callback = callback
-        self.slider_rep = None
-
-    def __call__(self, value, update_widget=False):
-        """Update the value."""
-        self.callback(value)
-        if update_widget:
-            if self.slider_rep is not None:
-                self.slider_rep.SetValue(value)
-                self.plotter.update()
 
 
 class _TimeViewer(object):
@@ -1323,159 +1038,6 @@ class _TimeViewer(object):
         self.picked_renderer = None
         for key in list(self.act_data_smooth.keys()):
             self.act_data_smooth[key] = None
-
-
-class _LinkViewer(object):
-    """Class to link multiple _TimeViewer objects."""
-
-    def __init__(self, brains, time=True, camera=False, colorbar=True,
-                 picking=False):
-        self.brains = brains
-        self.time_viewers = [brain.time_viewer for brain in brains]
-
-        # check time infos
-        times = [brain._times for brain in brains]
-        if time and not all(np.allclose(x, times[0]) for x in times):
-            warn('stc.times do not match, not linking time')
-            time = False
-
-        if camera:
-            self.link_cameras()
-
-        if time:
-            # link time sliders
-            self.link_sliders(
-                name="time",
-                callback=self.set_time_point,
-                event_type="always"
-            )
-
-            # link playback speed sliders
-            self.link_sliders(
-                name="playback_speed",
-                callback=self.set_playback_speed,
-                event_type="always"
-            )
-
-            # link toggle to start/pause playback
-            for time_viewer in self.time_viewers:
-                time_viewer.actions["play"].triggered.disconnect()
-                time_viewer.actions["play"].triggered.connect(
-                    self.toggle_playback)
-
-            # link time course canvas
-            def _time_func(*args, **kwargs):
-                for time_viewer in self.time_viewers:
-                    time_viewer.callbacks["time"](*args, **kwargs)
-
-            for time_viewer in self.time_viewers:
-                if time_viewer.show_traces:
-                    time_viewer.mpl_canvas.time_func = _time_func
-
-        if picking:
-            def _func_add(*args, **kwargs):
-                for time_viewer in self.time_viewers:
-                    time_viewer._add_point(*args, **kwargs)
-                    time_viewer.plotter.update()
-
-            def _func_remove(*args, **kwargs):
-                for time_viewer in self.time_viewers:
-                    time_viewer._remove_point(*args, **kwargs)
-
-            # save initial picked points
-            initial_points = dict()
-            for hemi in ('lh', 'rh'):
-                initial_points[hemi] = set()
-                for time_viewer in self.time_viewers:
-                    initial_points[hemi] |= \
-                        set(time_viewer.picked_points[hemi])
-
-            # link the viewers
-            for time_viewer in self.time_viewers:
-                time_viewer.clear_points()
-                time_viewer._add_point = time_viewer.add_point
-                time_viewer.add_point = _func_add
-                time_viewer._remove_point = time_viewer.remove_point
-                time_viewer.remove_point = _func_remove
-
-            # link the initial points
-            leader = self.time_viewers[0]  # select a time_viewer as leader
-            for hemi in initial_points.keys():
-                if hemi in time_viewer.brain._hemi_meshes:
-                    mesh = time_viewer.brain._hemi_meshes[hemi]
-                    for vertex_id in initial_points[hemi]:
-                        leader.add_point(hemi, mesh, vertex_id)
-
-        if colorbar:
-            leader = self.time_viewers[0]  # select a time_viewer as leader
-            fmin = leader.brain._data["fmin"]
-            fmid = leader.brain._data["fmid"]
-            fmax = leader.brain._data["fmax"]
-            for time_viewer in self.time_viewers:
-                time_viewer.callbacks["fmin"](fmin)
-                time_viewer.callbacks["fmid"](fmid)
-                time_viewer.callbacks["fmax"](fmax)
-
-            for slider_name in ('fmin', 'fmid', 'fmax'):
-                func = getattr(self, "set_" + slider_name)
-                self.link_sliders(
-                    name=slider_name,
-                    callback=func,
-                    event_type="always"
-                )
-
-    def set_fmin(self, value):
-        for time_viewer in self.time_viewers:
-            time_viewer.callbacks["fmin"](value)
-
-    def set_fmid(self, value):
-        for time_viewer in self.time_viewers:
-            time_viewer.callbacks["fmid"](value)
-
-    def set_fmax(self, value):
-        for time_viewer in self.time_viewers:
-            time_viewer.callbacks["fmax"](value)
-
-    def set_time_point(self, value):
-        for time_viewer in self.time_viewers:
-            time_viewer.callbacks["time"](value, update_widget=True)
-
-    def set_playback_speed(self, value):
-        for time_viewer in self.time_viewers:
-            time_viewer.callbacks["playback_speed"](value, update_widget=True)
-
-    def toggle_playback(self):
-        leader = self.time_viewers[0]  # select a time_viewer as leader
-        value = leader.callbacks["time"].slider_rep.GetValue()
-        # synchronize starting points before playback
-        self.set_time_point(value)
-        for time_viewer in self.time_viewers:
-            time_viewer.toggle_playback()
-
-    def link_sliders(self, name, callback, event_type):
-        from ..backends._pyvista import _update_slider_callback
-        for time_viewer in self.time_viewers:
-            slider = time_viewer.sliders[name]
-            if slider is not None:
-                _update_slider_callback(
-                    slider=slider,
-                    callback=callback,
-                    event_type=event_type
-                )
-
-    def link_cameras(self):
-        from ..backends._pyvista import _add_camera_callback
-
-        def _update_camera(vtk_picker, event):
-            for time_viewer in self.time_viewers:
-                time_viewer.plotter.update()
-
-        leader = self.time_viewers[0]  # select a time_viewer as leader
-        camera = leader.plotter.camera
-        _add_camera_callback(camera, _update_camera)
-        for time_viewer in self.time_viewers:
-            for renderer in time_viewer.plotter.renderers:
-                renderer.camera = camera
 
 
 def _get_range(brain):
